Use some extra cleverness to pick the bite.
The goal here is that each day there should exactly one item
returned from (bite).  That item should be the same throughout
the day, even in the face of adding more data, and if I
refactored a bit to allow bite to take a date parameter in the
past, we could get the bite for that day, too.